sort files in chatbot
