+++ src/Pages/Chatbot/componentsChatbot/index.ts
@@ -1,0 +1,9 @@
+export { default as Orb } from './Orb';
+export { default as IntroAnimation } from './IntroAnimation';
+export { default as BackgroundDecorations } from './BackgroundDecorations';
+export { default as Sidebar } from './Sidebar';
+export { default as Header } from './Header';
+export { default as MessageBubble } from './MessageBubble';
+export { default as LoadingIndicator } from './LoadingIndicator';
+export { default as ChatInput } from './ChatInput';
+export * from './types';
